+++ src/main/clojure/com/github/clojure_lsp/intellij/extension/documentation.clj
@@ -4,7 +4,7 @@
    :implements [com.intellij.lang.documentation.DocumentationProvider])
   (:require
    [com.github.clojure-lsp.intellij.db :as db]
-   [com.github.clojure-lsp.intellij.lsp-client :as lsp-client]
+   [com.github.clojure-lsp.intellij.client :as lsp-client]
    [markdown.core :as markdown])
   (:import
    [com.intellij.openapi.util.text StringUtil]