+++ src/main/clojure/com/github/clojure_lsp/intellij/extension/general_settings.clj
@@ -6,7 +6,7 @@
    [clojure.pprint :as pprint]
    [clojure.walk :as walk]
    [com.github.clojure-lsp.intellij.db :as db]
-   [com.github.clojure-lsp.intellij.lsp-client :as lsp-client]
+   [com.github.clojure-lsp.intellij.client :as lsp-client]
    [seesaw.core :as see]
    [seesaw.mig :as mig])
   (:import