+++ src/main/clojure/com/github/clojure_lsp/intellij/server.clj
@@ -4,7 +4,7 @@
    [clojure.java.io :as io]
    [com.github.clojure-lsp.intellij.db :as db]
    [com.github.clojure-lsp.intellij.logger :as logger]
-   [com.github.clojure-lsp.intellij.lsp-client :as lsp-client]
+   [com.github.clojure-lsp.intellij.client :as lsp-client]
    [com.github.clojure-lsp.intellij.tasks :as tasks]
    [lsp4clj.server :as lsp4clj.server])
   (:import